Refactor kernel simulator.
Separate kernel simulator implementation from actual kernel api definition so it can be retargeted to different api.
